Added support for Firefox
https://bugzilla.mozilla.org/show_bug.cgi?id=1332942
Because of a bug in firefox it is impossible to change the first
line imediately after text input. For now we will just bump down
the first result to the second line and leave the first line as is.

Once the bug is fixed we can delete this branch and use master for
both browser releases.